Moves the currency update settings general handler to the new handlers.js file for these settings
